57151 - [64] 64 bit support needed for Windows XP 64-Bit Edition

+++ bundles/org.eclipse.swt/Eclipse SWT Drag and Drop/win32/org/eclipse/swt/dnd/TableDropTargetEffect.java
@@ -103,7 +103,7 @@ public class TableDropTargetEffect extends DropTargetEffect {
 	 */
 	public void dragLeave(DropTargetEvent event) {
 		Table table = (Table) control;
-		int handle = table.handle;
+		int /*long*/ handle = table.handle;
 		if (dropHighlight != null) {
 			LVITEM lvItem = new LVITEM ();
 			lvItem.stateMask = OS.LVIS_DROPHILITED;
@@ -134,7 +134,7 @@ public class TableDropTargetEffect extends DropTargetEffect {
 	public void dragOver(DropTargetEvent event) {
 		Table table = (Table) getControl();
 		int effect = checkEffect(event.feedback);
-		int handle = table.handle;
+		int /*long*/ handle = table.handle;
 		Point coordinates = new Point(event.x, event.y);
 		coordinates = table.toControl(coordinates);
 		LVHITTESTINFO pinfo = new LVHITTESTINFO();
@@ -147,8 +147,8 @@ public class TableDropTargetEffect extends DropTargetEffect {
 		} else {
 			if (pinfo.iItem != -1 && scrollIndex == pinfo.iItem && scrollBeginTime != 0) {
 				if (System.currentTimeMillis() >= scrollBeginTime) {
-					int top = Math.max (0, OS.SendMessage (handle, OS.LVM_GETTOPINDEX, 0, 0));
-					int count = OS.SendMessage (handle, OS.LVM_GETITEMCOUNT, 0, 0);
+					int top = Math.max (0, (int)/*64*/OS.SendMessage (handle, OS.LVM_GETTOPINDEX, 0, 0));
+					int count = (int)/*64*/OS.SendMessage (handle, OS.LVM_GETITEMCOUNT, 0, 0);
 					int index = (scrollIndex - 1 < top) ? Math.max(0, scrollIndex - 1) : Math.min(count - 1, scrollIndex + 1);
 					boolean scroll = true;
 					if (pinfo.iItem == top) {
