+++ bundles/org.eclipse.swt/Eclipse SWT/win32/org/eclipse/swt/widgets/DirectoryDialog.java
@@ -80,8 +80,8 @@ public DirectoryDialog (Shell parent, int style) {
 	checkSubclass ();
 }
 
-int BrowseCallbackProc (int hwnd, int uMsg, int lParam, int lpData) {
-	switch (uMsg) {
+int /*long*/ BrowseCallbackProc (int /*long*/ hwnd, int /*long*/ uMsg, int /*long*/ lParam, int /*long*/ lpData) {
+	switch ((int)/*64*/uMsg) {
 		case OS.BFFM_INITIALIZED:
 			if (filterPath != null && filterPath.length () != 0) {
 				/* Use the character encoding for the default locale */
@@ -145,14 +145,14 @@ public String getMessage () {
 public String open () {
 	if (OS.IsWinCE) SWT.error (SWT.ERROR_NOT_IMPLEMENTED);
 	
-	int hHeap = OS.GetProcessHeap ();
+	int /*long*/ hHeap = OS.GetProcessHeap ();
 	
 	/* Get the owner HWND for the dialog */
-	int hwndOwner = 0;
+	int /*long*/ hwndOwner = 0;
 	if (parent != null) hwndOwner = parent.handle;
 
 	/* Copy the message to OS memory */
-	int lpszTitle = 0;
+	int /*long*/ lpszTitle = 0;
 	if (message.length () != 0) {
 		String string = message;
 		if (string.indexOf ('&') != -1) {
@@ -175,7 +175,7 @@ public String open () {
 
 	/* Create the BrowseCallbackProc */
 	Callback callback = new Callback (this, "BrowseCallbackProc", 4); //$NON-NLS-1$
-	int lpfn = callback.getAddress ();
+	int /*long*/ lpfn = callback.getAddress ();
 	if (lpfn == 0) SWT.error (SWT.ERROR_NO_MORE_CALLBACKS);
 	
 	/* Make the parent shell be temporary modal */
@@ -224,7 +224,7 @@ public String open () {
 	*/
 	boolean oldRunMessages = display.runMessages;
 	if (OS.COMCTL32_MAJOR < 6) display.runMessages = false;
-	int lpItemIdList = OS.SHBrowseForFolder (lpbi);
+	int /*long*/ lpItemIdList = OS.SHBrowseForFolder (lpbi);
 	if (OS.COMCTL32_MAJOR < 6) display.runMessages = oldRunMessages;
 	OS.SetErrorMode (oldErrorMode);
 	
@@ -250,7 +250,7 @@ public String open () {
 	if (lpszTitle != 0) OS.HeapFree (hHeap, 0, lpszTitle);
 
 	/* Free the pointer to the ITEMIDLIST */
-	int [] ppMalloc = new int [1];
+	int /*long*/ [] ppMalloc = new int /*long*/ [1];
 	if (OS.SHGetMalloc (ppMalloc) == OS.S_OK) {
 		/* void Free (struct IMalloc *this, void *pv); */
 		OS.VtblCall (5, ppMalloc [0], lpItemIdList);
